refactor: extract organization skills pane
